now using command line args
