+++ README.md
@@ -1,1 +1,3 @@
 # Cowtown-2022-Scouting-Compiler
+
+the command i used to run this was `gradlew run --args="path/to/1.csv path/to/2.csv"`
